requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==19.3.0
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
discord.py==1.3.3
idna==2.9
more-itertools==8.3.0
multidict==4.7.6
packaging==20.4
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.2
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.9
wcwidth==0.1.9
websockets==8.1
yarl==1.4.2
